requirements.txt
# packages for disease tracker tutorial
cssselect
pdfminer
PyPDF2
requests

# Packages for this repo
apache-airflow==2.2.4  # constraint = python 3.8
apache-airflow-providers-mongo==2.3.0

# Dependencies
alembic==1.7.6
anyio==3.5.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.3
apache-airflow-providers-imap==2.2.0
apache-airflow-providers-sqlite==2.1.0
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
blinker==1.4
cachelib==0.6.0
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.11.1
croniter==1.3.4
# I had to `pip install -U cryptography` on my M1 mac.
# ended up with cryptography==36.0.1
# cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.0
docutils==0.16
email-validator==1.1.3
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
Flask==1.1.2
graphviz==0.19.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.0.2
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==3.0.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
marshmallow==3.14.1
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
packaging==21.3
pendulum==2.1.2
pip==22.0.4
prison==0.2.1
psutil==5.9.0
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
pymongo==3.12.3
pyparsing==2.4.7
pyrsistent==0.16.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
rfc3986==1.5.0
rich==11.2.0
setproctitle==1.2.2
setuptools==60.9.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
SQLAlchemy==1.3.24
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
unicodecsv==0.14.1
urllib3==1.26.8
Werkzeug==1.0.1
wheel==0.37.1
wrapt==1.13.3
WTForms==2.3.3
zipp==3.7.0
